+++ .pre-commit-config.yaml
@@ -1,11 +1,11 @@
 repos:
-    - repo: https://github.com/psf/black
-      rev: 19.10b0
-      hooks:
-        - id: black
-    - repo: https://github.com/pre-commit/pre-commit-hooks
-      rev: v3.1.0
-      hooks:
-        - id: check-byte-order-marker
-        - id: trailing-whitespace
-        - id: end-of-file-fixer
+  - repo: https://github.com/psf/black
+    rev: 19.10b0
+    hooks:
+      - id: black
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v3.1.0
+    hooks:
+      - id: check-byte-order-marker
+      - id: trailing-whitespace
+      - id: end-of-file-fixer
