+++ readme.html
@@ -1,10 +1,10 @@
 <html>
-  <head></head>
-  <body>
-    <h1>Code42 Phantom App</h1>
-    <p>This app implements various investigative actions on the Code42 Incydr platform.</p>
-    <h2>How to Configure the App</h2>
-    <p>Access the Asset Settings tab on the Asset Configuration page. Input the cloud instance, username, and password
-      to use to connect to Code42.</p>
-  </body>
+<head></head>
+<body>
+<h1>Code42 Phantom App</h1>
+<p>This app implements various investigative actions on the Code42 Incydr platform.</p>
+<h2>How to Configure the App</h2>
+<p>Access the Asset Settings tab on the Asset Configuration page. Input the cloud instance, username, and password
+    to use to connect to Code42.</p>
+</body>
 </html>
